Recommend Plugin for Multi Monitor Setups
Having multiple monitors combined with GNOME's global hotkeys for
switching workspaces or moving windows to different workspaces leads to
the fact that the window on the focussed workspace is not foccused.
Thus, it is not possible to type into the window (see also #704).
Combining Focus Follows Workspace and PaperWM fixes this issues and
provides a nice multi monitor workflow.

Fixes #704